Onyx: Shit, added resources.rc and icon.ico. A warning Cleanup.

+++ Branch/Onyx 2.4.3/src/onyx-world/AIInterface.cpp
@@ -1907,7 +1907,7 @@ void AIInterface::SendMoveToPacket(float toX, float toY, float toZ, float toO, u
 	for(set<Player*>::iterator itr = m_Unit->GetInRangePlayerSetBegin(); itr != m_Unit->GetInRangePlayerSetEnd(); ++itr)
 	{
 		if( (*itr)->GetPositionNC().Distance2DSq( m_Unit->GetPosition() ) >= World::m_movementCompressThresholdCreatures )
-			(*itr)->AppendMovementData( SMSG_MONSTER_MOVE, data.GetSize(), (const uint8*)data.GetBufferPointer() );
+			(*itr)->AppendMovementData( SMSG_MONSTER_MOVE, (uint32)data.GetSize(), (const uint8*)data.GetBufferPointer() );
 		else
 			(*itr)->GetSession()->SendPacket(&data);
 	}
